Issue9_destroyアクションと価格修正

+++ app/controllers/items_controller.rb
@@ -39,7 +39,9 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
+      if item.user_id == current_user.id
+        item.destroy
+      end
     redirect_to root_path
   end
